Put items collection into seperate users

+++ lib/repositories/items_repository.dart
@@ -1,9 +1,16 @@
 import 'package:cantwait28/models/item_model.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
+import 'package:firebase_auth/firebase_auth.dart';
 
 class ItemsRepository {
 Stream<List<ItemModel>> getItemsStream () {
+  final userID = FirebaseAuth.instance.currentUser?.uid;
+  if (userID == null) {
+    throw Exception('User is not logged in');
+  }
   return FirebaseFirestore.instance
+        .collection('users')
+        .doc(userID)
         .collection('items')
         .orderBy('release_date')
         .snapshots()
@@ -19,13 +26,25 @@ Stream<List<ItemModel>> getItemsStream () {
 }
 
 Future<void> delete({required String id}) {
+  final userID = FirebaseAuth.instance.currentUser?.uid;
+  if (userID == null) {
+    throw Exception('User is not logged in');
+  }
   return FirebaseFirestore.instance
+          .collection('users')
+          .doc(userID)
           .collection('items')
           .doc(id)
           .delete();
 }
 Future<ItemModel> get({required String id}) async {
+  final userID = FirebaseAuth.instance.currentUser?.uid;
+  if (userID == null) {
+    throw Exception('User is not logged in');
+  }
   final doc = await FirebaseFirestore.instance
+          .collection('users')
+          .doc(userID)
           .collection('items')
           .doc(id)
           .get();
@@ -37,8 +56,15 @@ Future<ItemModel> get({required String id}) async {
           );
 }
 Future<void> add(String title, String imageURL, DateTime releaseDate,) async {
+  final userID = FirebaseAuth.instance.currentUser?.uid;
+  if (userID == null) {
+    throw Exception('User is not logged in');
+  }
      {
-      await FirebaseFirestore.instance.collection('items').add(
+      await FirebaseFirestore.instance.collection('users')
+      .doc(userID)
+      .collection('items')
+      .add(
         {
           'title': title,
           'image_url': imageURL,
